Reorganize exports
The most important stuff is now at top level, the rest is in

index
query
store
