+++ banyan-utils/src/bin/cli.rs
@@ -7,12 +7,9 @@ use structopt::StructOpt;
 use tracing::Level;
 
 use banyan::{
-    forest::*,
-    memstore::MemStore,
     query::{AllQuery, OffsetRangeQuery, QueryExt},
-    store::{ArcBlockWriter, ArcReadOnlyStore, BlockWriter, ReadOnlyStore},
-    tree::*,
-    StreamBuilder,
+    store::{ArcBlockWriter, ArcReadOnlyStore, BlockWriter, MemStore, ReadOnlyStore},
+    BranchCache, Config, Forest, Secrets, StreamBuilder, Transaction, Tree,
 };
 use banyan_utils::{
     create_chacha_key, dump,
