+++ banyan/tests/link_scraping.rs
@@ -1,10 +1,8 @@
 #![allow(clippy::upper_case_acronyms)]
 use banyan::{
-    forest::{BranchCache, Forest, Transaction, TreeTypes},
     index::{UnitSeq, VecSeq},
-    memstore::MemStore,
-    tree::Tree,
-    StreamBuilder,
+    store::MemStore,
+    BranchCache, Forest, StreamBuilder, Transaction, Tree, TreeTypes,
 };
 use common::Sha256Digest;
 use fnv::FnvHashSet;
